+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Next
 
+* Remove deprecated AAD graph code
 * Remove partner ID from user agent string [[GH-95](https://github.com/hashicorp/vault-plugin-secrets-azure/pull/95)]
 * Fixes environment not being used when using MS Graph [[GH-87](https://github.com/hashicorp/vault-plugin-secrets-azure/pull/87)]